requirements.txt
llama-cpp-python==0.2.11
numpy>=1.24.0
flask==2.2.5
flask-restx==1.1.0
werkzeug==2.2.3
requests>=2.25.0 